requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
Pillow>=10.0.0
numpy
opencv-python
torch  # PyTorch 추가
pandas  # YOLOv5 결과 처리를 위한 pandas 추가
requests  # YOLOv5에서 사용하는 requests 추가
matplotlib  # 선택 사항, 필요시 주석 제거
